Updated from suggestions by Smeech
Cleaned it up a bit, but it is still kind of a beast.  I probably need to optimize the variables a bit, since I basically have two of the same prompt.  The url prompt creates a URL for prompting, so it uses %0A to add the line break.  The boxprompt gets pasted in as text and it uses \n for newline.  The box prompt also doesn't define the model, so that is not in that one.  

urlprompt uses :oasis
boxpromt uses :poasis